requirements.txt
# Framework Web
fastapi>=0.100.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database e Cache
redis>=4.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6
aioredis>=2.0.0

# AI e ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.0.0
diffusers>=0.19.0
accelerate>=0.21.0
safetensors>=0.3.1

# Processamento de Mídia
Pillow>=9.0.0
numpy>=1.21.0
opencv-python-headless>=4.5.0
ffmpeg-python>=0.2.0

# Monitoramento e Logging
prometheus-client>=0.14.0
python-json-logger>=2.0.7
opentelemetry-api>=1.0.0
opentelemetry-sdk>=1.0.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation>=0.41b0
opentelemetry-exporter-otlp>=1.0.0
psutil>=5.9.0

# Utilitários
python-dotenv>=1.0.0
pydantic>=2.0.0
email-validator>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
aiohttp>=3.8.0
tenacity>=8.2.2
itsdangerous>=2.0.0
semver>=3.0.0
PyYAML>=6.0.0
gradio>=4.19.1
colorama>=0.4.6
python-slugify>=8.0.0
typing-extensions>=4.5.0

# Processamento de texto
freetype-py>=2.3.0
# PyICU>=2.12.0  # Requer: apt-get install pkg-config libicu-dev python3-dev
PyICU>=2.12.0; platform_system != "Windows"  # Só instalar em sistemas Unix-like

# Processamento de vídeo
moviepy>=1.0.3
pydub>=0.25.1
soundfile>=0.12.1

# Processamento de mídia
einops>=0.6.0
pytorch-lightning>=2.0.9
aiofiles>=23.0.0

# Dependências de desenvolvimento
pytest>=7.4.2
pytest-asyncio>=0.21.1
black>=23.9.1
flake8>=6.1.0
mypy>=1.5.1

# Dependências de cache/queue
celery>=5.3.4

# Agendamento e tarefas
APScheduler>=3.10.1

# Rate Limiting
slowapi>=0.1.8

# Adicionadas novas dependências
websockets>=10.0
aiohttp>=3.8.0
yt-dlp>=2023.11.16